fix potential overflow

+++ eip1962cpp/src/api.cpp
@@ -79,7 +79,7 @@ std::vector<std::uint8_t> run_operation_extension(u8 operation, u8 mod_byte_len,
         }
 
         // Check if remaining input size is exact
-        u32 const expected_pair_len = 2 * extension_degree * mod_byte_len + wc.order_len();
+        u32 const expected_pair_len = u32(2) * u32(extension_degree) * u32(mod_byte_len) + u32(wc.order_len());
         if (deserializer.remaining() != expected_pair_len)
         {
             input_err("Input length is invalid for number of pairs");
